fix: move activity card test helpers to src/testing
Relocate harness and mocks out of workflow-activity/components so
feature-boundary lint passes after renaming helpers off .test filenames.
